asfdas

+++ src/main/java/com/example/ABCBankDetails/repository/AccountPayeeRepository.java
@@ -8,14 +8,7 @@ import com.example.ABCBankDetails.model.*;
 @Repository
 public interface AccountPayeeRepository extends JpaRepository<AccountPayeeRepository, Long>{
 
-	List<AccountPayee> findByAccountNumber(Long accountNumber);
-
-
-
-
-	
-	
-
+	List<AccountPayee> findByAccountNumber(int accountNumber);
 void deleteByAccountNoSourceAndAccountNoTarget(int accountNoSource,int accountNoTarget);
